% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile.R
\name{get_profile}
\alias{get_profile}
\title{Get Google Scholar Profile}
\usage{
get_profile(sortby_date = FALSE)
}
\arguments{
\item{sortby_date}{Logical. If \code{TRUE}, the publications are sorted by date.}

\item{id}{Your google scholar identifier. You can find it in the URL of your google scholar profile.}
}
\value{
a \code{Profile} object with list structure.
}
\description{
Get Google Scholar Profile
}
\examples{
r <- get_profile("FvNp0NkAAAAJ")
r
}
