% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyscholar.R
\name{tinyscholar}
\alias{tinyscholar}
\title{Get Google Scholar Profile}
\usage{
tinyscholar(id, sortby_date = FALSE, server = c("hiplot", "cse"))
}
\arguments{
\item{id}{Your google scholar identifier. You can find it in the URL of your google scholar profile.}

\item{sortby_date}{Logical. If \code{TRUE}, the publications are sorted by date.}

\item{server}{Web api server, default is \href{https://hiplot.com.cn/}{hiplot platform}.}
}
\value{
a \code{Profile} object with list structure.
}
\description{
Get Google Scholar Profile
}
\examples{
r <- tinyscholar("FvNp0NkAAAAJ")
r
tb <- scholar_table(r)
tb$citations
tb$publications
pl <- scholar_plot(r)
pl$citations
pl$publications
}
