% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scholar_plot.R
\name{scholar_plot}
\alias{scholar_plot}
\title{Show Scholar Profile Plot.}
\usage{
scholar_plot(
  profile,
  bar_width = 0.5,
  add_text = TRUE,
  title_citations = NULL,
  title_publications = NULL,
  caption_citations = paste("Update:", Sys.Date()),
  caption_publications = caption_citations
)
}
\arguments{
\item{profile}{Result from \link{tinyscholar}.}

\item{bar_width}{bar width.}

\item{add_text}{If \code{TRUE} (default), add text on the top of bar.}

\item{title_citations}{Title for plot \code{citations}. Set by \link[ggplot2:labs]{ggplot2::labs}.}

\item{title_publications}{Title for plot \code{publications}. Set by \link[ggplot2:labs]{ggplot2::labs}.}

\item{caption_citations}{Caption for plot \code{citations}. Set by \link[ggplot2:labs]{ggplot2::labs}.}

\item{caption_publications}{Caption for plot \code{publications}. Set by \link[ggplot2:labs]{ggplot2::labs}.}
}
\value{
a length-2 list of \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Show Scholar Profile Plot.
}
\seealso{
\link{tinyscholar}, \link{scholar_table}
}
