% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyscholar-package.R
\docType{package}
\name{tinyscholar-package}
\alias{tinyscholar}
\alias{tinyscholar-package}
\title{tinyscholar: Get and Show Google Scholar Profile}
\description{
A tiny package to get google scholar profile data
    from web api and show it in table or figure format.
}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
